+++ lib/main.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:pin_code_fields/pin_code_fields.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'dart:async';
+import 'screens/splash_screen.dart';
 
 void main() {
   runApp(const MainApp());
@@ -17,376 +16,21 @@ class MainApp extends StatelessWidget {
       theme: ThemeData(
         useMaterial3: true,
         colorScheme: ColorScheme.light(
-          primary: Color(0xFF1E3D59), // Deep blue
-          secondary: Color(0xFF17B978), // Light blue-grey
+          primary: const Color(0xFF1E3D59),
+          secondary: const Color(0xFF17B978),
           surface: Colors.white,
           onPrimary: Colors.white,
           onSecondary: Colors.white,
-          onSurface: Color(0xFF1E3D59),
+          onSurface: const Color(0xFF1E3D59),
         ),
         textTheme: GoogleFonts.interTextTheme(),
-        appBarTheme: AppBarTheme(
+        appBarTheme: const AppBarTheme(
           backgroundColor: Color(0xFF1E3D59),
           foregroundColor: Colors.white,
           elevation: 0,
         ),
       ),
       home: const SplashScreen(),
-    );
-  }
-}
-
-class SplashScreen extends StatefulWidget {
-  const SplashScreen({super.key});
-
-  @override
-  State<SplashScreen> createState() => _SplashScreenState();
-}
-
-class _SplashScreenState extends State<SplashScreen>
-    with SingleTickerProviderStateMixin {
-  late AnimationController _controller;
-  late Animation<double> _fadeAnimation;
-  late Animation<double> _scaleAnimation;
-  late Animation<double> _rotateAnimation;
-  late Animation<double> _pulseAnimation;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller = AnimationController(
-      duration: const Duration(seconds: 2),
-      vsync: this,
-    );
-
-    _fadeAnimation = Tween<double>(begin: 0.0, end: 1.0).animate(
-      CurvedAnimation(
-        parent: _controller,
-        curve: Interval(0.0, 0.5, curve: Curves.easeIn),
-      ),
-    );
-
-    _scaleAnimation = Tween<double>(begin: 0.5, end: 1.0).animate(
-      CurvedAnimation(
-        parent: _controller,
-        curve: Interval(0.0, 0.5, curve: Curves.elasticOut),
-      ),
-    );
-
-    _rotateAnimation = Tween<double>(begin: -0.5, end: 0.0).animate(
-      CurvedAnimation(
-        parent: _controller,
-        curve: Interval(0.0, 0.5, curve: Curves.easeOut),
-      ),
-    );
-
-    _pulseAnimation = TweenSequence<double>([
-      TweenSequenceItem(
-        tween: Tween<double>(begin: 1.0, end: 1.2),
-        weight: 1.0,
-      ),
-      TweenSequenceItem(
-        tween: Tween<double>(begin: 1.2, end: 1.0),
-        weight: 1.0,
-      ),
-    ]).animate(
-      CurvedAnimation(
-        parent: _controller,
-        curve: Interval(0.5, 1.0, curve: Curves.easeInOut),
-      ),
-    );
-
-    _controller.repeat(reverse: true);
-
-    Timer(const Duration(seconds: 3), () {
-      _controller.stop();
-      Navigator.pushReplacement(
-        context,
-        PageRouteBuilder(
-          pageBuilder:
-              (context, animation, secondaryAnimation) => PinEntryScreen(),
-          transitionsBuilder: (context, animation, secondaryAnimation, child) {
-            return FadeTransition(opacity: animation, child: child);
-          },
-          transitionDuration: Duration(milliseconds: 800),
-        ),
-      );
-    });
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Theme.of(context).colorScheme.primary,
-      body: Center(
-        child: AnimatedBuilder(
-          animation: _controller,
-          builder: (context, child) {
-            return FadeTransition(
-              opacity: _fadeAnimation,
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  Transform.rotate(
-                    angle: _rotateAnimation.value * 3.14,
-                    child: ScaleTransition(
-                      scale: _scaleAnimation,
-                      child: Transform.scale(
-                        scale: _pulseAnimation.value,
-                        child: Container(
-                          padding: EdgeInsets.all(20),
-                          decoration: BoxDecoration(
-                            color: Colors.white,
-                            shape: BoxShape.circle,
-                            boxShadow: [
-                              BoxShadow(
-                                color: Colors.black12,
-                                blurRadius: 20,
-                                spreadRadius: 5,
-                              ),
-                            ],
-                          ),
-                          child: Icon(
-                            Icons.receipt_long,
-                            size: 60,
-                            color: Theme.of(context).colorScheme.primary,
-                          ),
-                        ),
-                      ),
-                    ),
-                  ),
-                  SizedBox(height: 30),
-                  ScaleTransition(
-                    scale: _scaleAnimation,
-                    child: Text(
-                      'BillBro',
-                      style: GoogleFonts.inter(
-                        fontSize: 32,
-                        fontWeight: FontWeight.bold,
-                        color: Colors.white,
-                      ),
-                    ),
-                  ),
-                  SizedBox(height: 8),
-                  FadeTransition(
-                    opacity: _fadeAnimation,
-                    child: Text(
-                      'Manage your business with ease',
-                      style: GoogleFonts.inter(
-                        fontSize: 16,
-                        color: Colors.white.withOpacity(0.8),
-                      ),
-                    ),
-                  ),
-                  SizedBox(height: 40),
-                  // Animated dots
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: List.generate(3, (index) {
-                      return Padding(
-                        padding: EdgeInsets.symmetric(horizontal: 4),
-                        child: ScaleTransition(
-                          scale: CurvedAnimation(
-                            parent: _controller,
-                            curve: Interval(
-                              index * 0.2,
-                              0.6 + index * 0.2,
-                              curve: Curves.easeInOut,
-                            ),
-                          ),
-                          child: Container(
-                            width: 8,
-                            height: 8,
-                            decoration: BoxDecoration(
-                              color: Theme.of(context).colorScheme.secondary,
-                              shape: BoxShape.circle,
-                            ),
-                          ),
-                        ),
-                      );
-                    }),
-                  ),
-                ],
-              ),
-            );
-          },
-        ),
-      ),
-    );
-  }
-}
-
-class PinEntryScreen extends StatefulWidget {
-  const PinEntryScreen({super.key});
-
-  @override
-  _PinEntryScreenState createState() => _PinEntryScreenState();
-}
-
-class _PinEntryScreenState extends State<PinEntryScreen> {
-  final TextEditingController pinController = TextEditingController();
-  final String correctPin = "1234"; // Change this as needed
-
-  void validatePin() {
-    if (pinController.text == correctPin) {
-      Navigator.push(
-        context,
-        MaterialPageRoute(builder: (context) => const HomePage()),
-      );
-    } else {
-      pinController.clear();
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(
-          content: Text('Incorrect PIN! Try again'),
-          backgroundColor: Colors.red,
-        ),
-      );
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Theme.of(context).colorScheme.surface,
-      body: SafeArea(
-        child: Padding(
-          padding: const EdgeInsets.all(20.0),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              const SizedBox(height: 40),
-              Text(
-                'Welcome Back',
-                style: GoogleFonts.inter(
-                  fontSize: 32,
-                  fontWeight: FontWeight.bold,
-                  color: Theme.of(context).colorScheme.primary,
-                ),
-              ),
-              const SizedBox(height: 8),
-              Text(
-                'Enter your PIN to access BillBro',
-                style: GoogleFonts.inter(fontSize: 16, color: Colors.grey[600]),
-              ),
-              const SizedBox(height: 60),
-              Center(
-                child: Column(
-                  children: [
-                    Container(
-                      padding: EdgeInsets.all(30),
-                      decoration: BoxDecoration(
-                        color: Colors.white,
-                        borderRadius: BorderRadius.circular(20),
-                        boxShadow: [
-                          BoxShadow(
-                            color: Colors.black.withOpacity(0.05),
-                            blurRadius: 20,
-                            offset: Offset(0, 10),
-                          ),
-                        ],
-                      ),
-                      child: Column(
-                        children: [
-                          PinCodeTextField(
-                            appContext: context,
-                            length: 4,
-                            obscureText: true,
-                            animationType: AnimationType.fade,
-                            keyboardType: TextInputType.number,
-                            controller: pinController,
-                            pinTheme: PinTheme(
-                              shape: PinCodeFieldShape.box,
-                              borderRadius: BorderRadius.circular(12),
-                              fieldHeight: 50,
-                              fieldWidth: 45,
-                              activeFillColor:
-                                  Theme.of(context).colorScheme.surface,
-                              selectedFillColor: Theme.of(
-                                context,
-                              ).colorScheme.primary.withOpacity(0.1),
-                              inactiveFillColor:
-                                  Theme.of(context).colorScheme.surface,
-                              activeColor:
-                                  Theme.of(context).colorScheme.primary,
-                              selectedColor:
-                                  Theme.of(context).colorScheme.primary,
-                              inactiveColor: Colors.grey[300],
-                            ),
-                            enableActiveFill: true,
-                            onChanged: (value) {},
-                          ),
-                          const SizedBox(height: 30),
-                          ElevatedButton(
-                            onPressed: validatePin,
-                            style: ElevatedButton.styleFrom(
-                              backgroundColor:
-                                  Theme.of(context).colorScheme.primary,
-                              foregroundColor: Colors.white,
-                              padding: const EdgeInsets.symmetric(
-                                horizontal: 50,
-                                vertical: 15,
-                              ),
-                              shape: RoundedRectangleBorder(
-                                borderRadius: BorderRadius.circular(12),
-                              ),
-                              minimumSize: Size(double.infinity, 55),
-                            ),
-                            child: Text(
-                              "Unlock",
-                              style: GoogleFonts.inter(
-                                fontSize: 18,
-                                fontWeight: FontWeight.w600,
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class HomePage extends StatelessWidget {
-  const HomePage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Theme.of(context).colorScheme.surface,
-      appBar: AppBar(
-        title: Text(
-          'BillBro',
-          style: GoogleFonts.inter(fontWeight: FontWeight.w600),
-        ),
-        centerTitle: true,
-      ),
-      body: Center(
-        child: Text(
-          "Welcome to your BillBro Dashboard!",
-          style: GoogleFonts.inter(
-            fontSize: 20,
-            color: Theme.of(context).colorScheme.primary,
-          ),
-        ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {},
-        backgroundColor: Theme.of(context).colorScheme.secondary,
-        child: Icon(Icons.add, color: Colors.white),
-      ),
     );
   }
 }
