+++ lib/screens/pin_entry_screen.dart
@@ -1,0 +1,141 @@
+import 'package:flutter/material.dart';
+import 'package:pin_code_fields/pin_code_fields.dart';
+import 'package:google_fonts/google_fonts.dart';
+import 'home_screen.dart';
+
+class PinEntryScreen extends StatefulWidget {
+  const PinEntryScreen({super.key});
+
+  @override
+  _PinEntryScreenState createState() => _PinEntryScreenState();
+}
+
+class _PinEntryScreenState extends State<PinEntryScreen> {
+  final TextEditingController pinController = TextEditingController();
+  final String correctPin = "1234"; // Change this as needed
+
+  void validatePin() {
+    if (pinController.text == correctPin) {
+      Navigator.push(
+        context,
+        MaterialPageRoute(builder: (context) => const HomePage()),
+      );
+    } else {
+      pinController.clear();
+      ScaffoldMessenger.of(context).showSnackBar(
+        const SnackBar(
+          content: Text('Incorrect PIN! Try again'),
+          backgroundColor: Colors.red,
+        ),
+      );
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Theme.of(context).colorScheme.surface,
+      body: SafeArea(
+        child: Padding(
+          padding: const EdgeInsets.all(20.0),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              const SizedBox(height: 40),
+              Text(
+                'Welcome Back',
+                style: GoogleFonts.inter(
+                  fontSize: 32,
+                  fontWeight: FontWeight.bold,
+                  color: Theme.of(context).colorScheme.primary,
+                ),
+              ),
+              const SizedBox(height: 8),
+              Text(
+                'Enter your PIN to access BillBro',
+                style: GoogleFonts.inter(fontSize: 16, color: Colors.grey[600]),
+              ),
+              const SizedBox(height: 60),
+              Center(
+                child: Column(
+                  children: [
+                    Container(
+                      padding: EdgeInsets.all(30),
+                      decoration: BoxDecoration(
+                        color: Colors.white,
+                        borderRadius: BorderRadius.circular(20),
+                        boxShadow: [
+                          BoxShadow(
+                            color: Colors.black.withOpacity(0.05),
+                            blurRadius: 20,
+                            offset: Offset(0, 10),
+                          ),
+                        ],
+                      ),
+                      child: Column(
+                        children: [
+                          PinCodeTextField(
+                            appContext: context,
+                            length: 4,
+                            obscureText: true,
+                            animationType: AnimationType.fade,
+                            keyboardType: TextInputType.number,
+                            controller: pinController,
+                            pinTheme: PinTheme(
+                              shape: PinCodeFieldShape.box,
+                              borderRadius: BorderRadius.circular(12),
+                              fieldHeight: 50,
+                              fieldWidth: 45,
+                              activeFillColor:
+                                  Theme.of(context).colorScheme.surface,
+                              selectedFillColor: Theme.of(
+                                context,
+                              ).colorScheme.primary.withOpacity(0.1),
+                              inactiveFillColor:
+                                  Theme.of(context).colorScheme.surface,
+                              activeColor:
+                                  Theme.of(context).colorScheme.primary,
+                              selectedColor:
+                                  Theme.of(context).colorScheme.primary,
+                              inactiveColor: Colors.grey[300],
+                            ),
+                            enableActiveFill: true,
+                            onChanged: (value) {},
+                          ),
+                          const SizedBox(height: 30),
+                          ElevatedButton(
+                            onPressed: validatePin,
+                            style: ElevatedButton.styleFrom(
+                              backgroundColor:
+                                  Theme.of(context).colorScheme.primary,
+                              foregroundColor: Colors.white,
+                              padding: const EdgeInsets.symmetric(
+                                horizontal: 50,
+                                vertical: 15,
+                              ),
+                              shape: RoundedRectangleBorder(
+                                borderRadius: BorderRadius.circular(12),
+                              ),
+                              minimumSize: Size(double.infinity, 55),
+                            ),
+                            child: Text(
+                              "Unlock",
+                              style: GoogleFonts.inter(
+                                fontSize: 18,
+                                fontWeight: FontWeight.w600,
+                              ),
+                            ),
+                          ),
+                        ],
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
